FIX: Fix incorrect data type for id from int to long
DB use bigint for id

+++ Services/IHolidayService.cs
@@ -6,7 +6,7 @@ namespace lievee.Services
     {
         Task<ServiceResultNoData> CreateNewHoliday(DateOnly date, Users currentUser);
         Task<ServiceResult<List<Holiday>>> GetHolidays(DateOnly startDate, DateOnly endDate);
-        Task<ServiceResultNoData> UpdateHoliday(int holidayId, DateOnly newDate);
-        Task<ServiceResultNoData> DeleteHoliday(int holidayId);
+        Task<ServiceResultNoData> UpdateHoliday(long holidayId, DateOnly newDate);
+        Task<ServiceResultNoData> DeleteHoliday(long holidayId);
     }
 }
